+++ extensions/opencode/session-catalog.ts
@@ -6,6 +6,14 @@ import type {
   SessionsCatalogReadResult,
 } from "openclaw/plugin-sdk/session-catalog";
 import {
+  boundSessionCatalogTranscriptPage,
+  boundedSessionCatalogLimit,
+  decodeSessionCatalogCursor,
+  encodeSessionCatalogCursor,
+  isExactSessionCatalogCursor,
+  optionalSessionCatalogCursor,
+} from "openclaw/plugin-sdk/session-catalog-runtime";
+import {
   isRecord,
   normalizeBoundedOptionalString as optionalOpenCodeString,
 } from "openclaw/plugin-sdk/string-coerce-runtime";
@@ -17,13 +25,9 @@ import {
 
 const LOCAL_HOST_ID = "gateway";
 const DEFAULT_PAGE_LIMIT = 20;
-const MAX_PAGE_LIMIT = 100;
 const MAX_SEARCH_LENGTH = 500;
-const MAX_CURSOR_LENGTH = 128;
 const MAX_CLI_LIST_SESSIONS = 10_000;
 const MAX_CLI_OUTPUT_BYTES = 32 * 1024 * 1024;
-const MAX_TRANSCRIPT_ITEM_BYTES = 512 * 1024;
-const MAX_TRANSCRIPT_PAGE_BYTES = 20 * 1024 * 1024;
 const CLI_TIMEOUT_MS = 30_000;
 const OPENCODE_QUERY_CACHE_TTL_MS = 32_000;
 const OPENCODE_QUERY_CACHE_MAX_ENTRIES = 32;
@@ -96,115 +100,7 @@ type OpenCodeReadParams = {
   cursor?: string;
 };
 
-function boundedLimit(value: unknown, fallback = DEFAULT_PAGE_LIMIT): number {
-  if (value === undefined) {
-    return fallback;
-  }
-  if (!Number.isInteger(value) || Number(value) < 1 || Number(value) > MAX_PAGE_LIMIT) {
-    throw new Error(`limit must be an integer between 1 and ${String(MAX_PAGE_LIMIT)}`);
-  }
-  return Number(value);
-}
-
-function encodeCursor(offset: number): string {
-  return Buffer.from(JSON.stringify({ offset }), "utf8").toString("base64url");
-}
-
-function optionalRawCursor(value: unknown): string | undefined {
-  if (value === undefined) {
-    return undefined;
-  }
-  if (typeof value !== "string" || value.length === 0 || value.length > MAX_CURSOR_LENGTH) {
-    throw new Error("cursor is invalid");
-  }
-  return value;
-}
-
-function decodeCursor(value: unknown): number {
-  const cursor = optionalRawCursor(value);
-  if (cursor === undefined) {
-    return 0;
-  }
-  try {
-    const bytes = Buffer.from(cursor, "base64url");
-    if (bytes.toString("base64url") !== cursor) {
-      throw new Error("non-canonical base64url");
-    }
-    const parsed = JSON.parse(bytes.toString("utf8")) as unknown;
-    if (!isRecord(parsed) || !Number.isSafeInteger(parsed.offset) || Number(parsed.offset) < 0) {
-      throw new Error("invalid offset");
-    }
-    const offset = Number(parsed.offset);
-    if (encodeCursor(offset) !== cursor) {
-      throw new Error("non-canonical cursor payload");
-    }
-    return offset;
-  } catch (error) {
-    throw new Error("cursor is invalid", { cause: error });
-  }
-}
-
-export function isExactOpenCodeSessionCursor(value: unknown): value is string {
-  if (typeof value !== "string") {
-    return false;
-  }
-  try {
-    decodeCursor(value);
-    return true;
-  } catch {
-    return false;
-  }
-}
-
-function truncateUtf8(text: string, maxBytes: number): string {
-  if (Buffer.byteLength(text, "utf8") <= maxBytes) {
-    return text;
-  }
-  let low = 0;
-  let high = text.length;
-  while (low < high) {
-    const middle = Math.ceil((low + high) / 2);
-    if (Buffer.byteLength(text.slice(0, middle), "utf8") <= maxBytes - 3) {
-      low = middle;
-    } else {
-      high = middle - 1;
-    }
-  }
-  const end = low > 0 && /[\uD800-\uDBFF]/u.test(text.charAt(low - 1)) ? low - 1 : low;
-  return `${text.slice(0, end)}…`;
-}
-
-function transcriptPage(
-  items: SessionCatalogTranscriptItem[],
-  limit: number,
-  offset: number,
-): { items: SessionCatalogTranscriptItem[]; nextCursor?: string } {
-  const end = Math.max(0, items.length - offset);
-  const start = Math.max(0, end - limit);
-  const page: SessionCatalogTranscriptItem[] = [];
-  let pageBytes = 2;
-  for (let index = end - 1; index >= start; index -= 1) {
-    const item = items[index];
-    if (!item) {
-      continue;
-    }
-    const bounded: SessionCatalogTranscriptItem = {
-      ...item,
-      text: truncateUtf8(item.text ?? "", MAX_TRANSCRIPT_ITEM_BYTES),
-    };
-    const itemBytes = Buffer.byteLength(JSON.stringify(bounded), "utf8") + 1;
-    if (page.length > 0 && pageBytes + itemBytes > MAX_TRANSCRIPT_PAGE_BYTES) {
-      break;
-    }
-    page.unshift(bounded);
-    pageBytes += itemBytes;
-  }
-  const consumed = offset + page.length;
-  return {
-    items: page,
-    ...(consumed < items.length ? { nextCursor: encodeCursor(consumed) } : {}),
-  };
-}
+export const isExactOpenCodeSessionCursor = isExactSessionCatalogCursor;
 
 function parseListParams(
   value: unknown,
@@ -225,9 +121,9 @@ function parseListParams(
   if (value.searchTerm !== undefined && !searchTerm) {
     throw new Error("searchTerm is invalid");
   }
-  const cursor = optionalRawCursor(value.cursor);
+  const cursor = optionalSessionCatalogCursor(value.cursor);
   return {
-    limit: boundedLimit(value.limit),
+    limit: boundedSessionCatalogLimit(value.limit),
     ...(searchTerm ? { searchTerm } : {}),
     ...(cursor ? { cursor } : {}),
   };
@@ -247,10 +143,10 @@ function parseReadParams(
   if (!threadId || !SESSION_ID_PATTERN.test(threadId)) {
     throw new Error("threadId is invalid");
   }
-  const cursor = optionalRawCursor(value.cursor);
+  const cursor = optionalSessionCatalogCursor(value.cursor);
   return {
     threadId,
-    limit: boundedLimit(value.limit),
+    limit: boundedSessionCatalogLimit(value.limit),
     ...(cursor ? { cursor } : {}),
   };
 }
@@ -414,7 +310,7 @@ export async function listLocalOpenCodeSessionPage(
   options: OpenCodeQueryCacheOptions = {},
 ): Promise<OpenCodeSessionPage> {
   const params = parseListParams(value);
-  const offset = decodeCursor(params.cursor);
+  const offset = decodeSessionCatalogCursor(params.cursor);
   const requestedCount = params.searchTerm
     ? MAX_CLI_LIST_SESSIONS
     : Math.min(MAX_CLI_LIST_SESSIONS, offset + params.limit + 1);
@@ -446,7 +342,7 @@ export async function listLocalOpenCodeSessionPage(
   return {
     sessions: page,
     ...(offset + page.length < sessions.length
-      ? { nextCursor: encodeCursor(offset + page.length) }
+      ? { nextCursor: encodeSessionCatalogCursor(offset + page.length) }
       : {}),
   };
 }
@@ -557,9 +453,9 @@ export async function readLocalOpenCodeTranscriptPage(
   value: unknown,
 ): Promise<SessionsCatalogReadResult> {
   const params = parseReadParams(value);
-  const offset = decodeCursor(params.cursor);
+  const offset = decodeSessionCatalogCursor(params.cursor);
   const items = openCodeTranscriptItems(await exportOpenCodeSession(params.threadId));
-  const page = transcriptPage(items, params.limit, offset);
+  const page = boundSessionCatalogTranscriptPage(items, params.limit, offset);
   return {
     hostId: LOCAL_HOST_ID,
     label: "Local OpenCode",
